moving types to its separate file
