+++ packages/common/src/types/index.ts
@@ -4,6 +4,7 @@ export * from './avatars'
 export * from './bad-request'
 export * from './database'
 export * from './events'
+export * from './followings'
 export * from './not-found'
 export * from './transactions'
 export * from './users'
