+++ packages/common/src/types/users.ts
@@ -39,33 +39,3 @@ export const GetUserRelatedEventsReplySchema = Type.Object({
   attended: Type.Array(RelatedEventToUserSchema)
 })
 export type GetUserRelatedEventsReply = Static<typeof GetUserRelatedEventsReplySchema>
-
-export const PostFollowRequestParamsSchema = Type.Object({
-  username: Type.String()
-})
-export type PostFollowRequestParams = Static<typeof PostFollowRequestParamsSchema>
-
-export const PostFollowRequestBodySchema = Type.Object({
-  usernameToFollow: Type.String()
-})
-export type PostFollowRequestBody = Static<typeof PostFollowRequestBodySchema>
-
-export const PostFollowReplySchema = Type.Object({
-  message: Type.String()
-})
-export type PostFollowReply = Static<typeof PostFollowReplySchema>
-
-export const PostUnfollowRequestParamsSchema = Type.Object({
-  username: Type.String()
-})
-export type PostUnfollowRequestParams = Static<typeof PostUnfollowRequestParamsSchema>
-
-export const PostUnfollowRequestBodySchema = Type.Object({
-  usernameToUnfollow: Type.String()
-})
-export type PostUnfollowRequestBody = Static<typeof PostUnfollowRequestBodySchema>
-
-export const PostUnfollowReplySchema = Type.Object({
-  message: Type.String()
-})
-export type PostUnfollowReply = Static<typeof PostUnfollowReplySchema>
